+++ src/main/java/racingCar/App.java
@@ -1,10 +1,11 @@
 package racingCar;
 
-import racingCar.ui.InputView;
+import racingCar.model.RacingCarGame;
+import racingCar.view.InputView;
 
 public class App {
     public static void main(String[] args) {
         InputView inputView = new InputView();
-        new RacingCarGame(inputView.inputCarNames(), inputView.inputTryTimes()).playRace();
+//        new RacingCarGame(inputView.inputCarNames(), inputView.inputTryTimes());
     }
 }
