+++ src/main/java/racingCar/model/RacingCarGame.java
@@ -1,29 +1,22 @@
-package racingCar;
+package racingCar.model;
 
-import racingCar.ui.ResultView;
+import racingCar.model.Car;
+import racingCar.view.ResultView;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
 public class RacingCarGame {
     private int tryTimes;
     private List<Car> raceCar = new ArrayList<>();
     private ResultView resultView = new ResultView();
 
-    RacingCarGame(String[] carNames, int tryTimes) {
+    public RacingCarGame(String[] carNames, int tryTimes) {
         if (tryTimes < 0) {
             throw new IllegalArgumentException("올바른 수를 입력하세요.");
         }
         setRaceCar(carNames);
         this.tryTimes = tryTimes;
-    }
-
-    public void playRace() {
-        for (int i = 0; i < tryTimes; i++) {
-            resultView.printCarRace(raceCar);
-        }
-
-        RacingCarGameRule racingCarGameRule = new RacingCarGameRule(raceCar);
-        racingCarGameRule.printWinners();
     }
 
     private void setRaceCar(String[] carNames) {
@@ -34,6 +27,18 @@ public class RacingCarGame {
 
     public List<Car> getRaceCar() {
         return this.raceCar;
+    }
+
+    public List<Car> winnersCars() {
+        return raceCar.stream()
+                .filter(car -> car.getMoveCount() == maxMoveCar().getMoveCount())
+                .collect(Collectors.toList());
+    }
+
+    public Car maxMoveCar() {
+        return raceCar.stream()
+                .max(Comparator.comparingInt(Car::getMoveCount))
+                .orElseThrow(NoSuchElementException::new);
     }
 
 }
