+++ src/test/java/racingCar/CarTest.java
@@ -2,6 +2,8 @@ package racingCar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingCar.model.RacingCarGame;
+import racingCar.model.Car;
 
 import java.util.List;
